pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Agentic RAG Course Planning Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-groq>=0.1.0",
    "llama-parse>=0.5.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "langchain-google-genai>=4.2.1",
]
